commands/requirements.txt
requests
pyyaml
flask
mangum
asgiref
discord-py-interactions
pymongo
bson